feat(state): Update `Proposal` with `BlobID` (#7)
* Extend type `Proposal` with `BlobID` field

* Update `Proposal` methods `ValidateBasic()` and `String()` to use `BlobID`

* Add `BlobID` to `Proposal` protobuf.

Regenerated the proto files as well.
Moved the proto definition of `BlobID` to the `types.proto` file to fix an
import cycle.

* Add function to canonalize `BlobID`

* Update `Proposal` to/from protobuf functions to include `BlobID`

* Update `Proposal` unit tests to use `BlobID`

* Add method `IsNil()` to type `BlobID`

* `State.defaultDecideProposal() assigns an empty `BlobID` to `Proposal`.

Since `BlobID` is not used in the consensus logic yet, this change is harmless.
However, it allows us to keep this PR focused solely on `Proposal`-related
changes.

In a future PR, we’ll handle the proper creation of `BlobID` by implementing
blob splitting into multiple parts.

Additionally, this PR includes a refactor for improved readability.

* Add missing `BlobID` param to `decideProposal()`

* Add missing `BlobID` param in calls to `NewProposal()` in state_test.go

* Add missing `BlobID` param in calls to `NewProposal()` in replay_test.go

* Add missing `BlobID` param in calls to `NewProposal()` in pbts_test.go

* Add missing `BlobID` param in calls to `NewProposal()` in byzantine_test.go

* Fixed unit test TestConsMsgsVectors.

The `Proposal` sub-test was using an incorrect hex string to check the test's
results.
The string changed because we added the `BlobID` field to the `Proposal` type,
thus changing its encoding to protobuf.

Also rearrenged the code for better readability.

* Fixed unit test `TestWALEncoder`.

It was using an incorrect hex string to check the test's results.
The string changed because we added the `BlobID` field to the `Proposal` type,
thus changing its encoding to protobuf.

Also rearrenged the code for better readability.

* Fixed unit test `TestPrivvalVectors`.

The sub-tests "Proposal Request", "Proposal Response", and "Proposal Response
with error" were was using an incorrect hex string to check the test's results.

The string changed because we added the `BlobID` field to the `Proposal` type,
thus changing its encoding to protobuf.

Also rearrenged the code for better readability.

* Extend `EventDataCompleteProposal` with `BlobID` field

* `Hash()` method of type `Blob` checks return `nil` if empty blob.

* updated docs

* Update `State.isProposalComplete()`.

Rationale: If a block has an associated blob, we set the `BlobID` in `Proposal`
after returning from `PrepareProposal()`, causing the `IsNil()` check to
return false. If the block does not have a blob, the `BlobID` field in
`Proposal` will remain its zero value, making `IsNil()` return true.

If a blob is present, `cs.ProposalBlob.IsNil()` will always return true until
`cs.ProposalBlob` is assigned a non-empty `[]byte` slice. This assignment
happens only after we have received all the blob parts and we have successfully
reconstructed the blob.