feat(state): Add `BlobPartMessage` and `HasBlobPartMessage` to consensus state and reactor (#9)
* Add `BlobPartMessage` in consensus reactor

* Add `BlobPartMessage` case in consensus state `handleMsg()`

* add blob part in protobuf messages

* Add `HasProposalBlobPartMessage` to consensus reactor

* Implement `Wrap()` method for types `BlobPart` and `HasProposalBlobPart`

* Added `BlobPartMessage` and `HasBlobPartMessage` to functions
serializing/deserializing protobuf messages to/from their corresponding
Go types

* Add `ProposalBlobPart` to events

* Add `BlobParts` to `peerStateStats`.

Does it make sense to mark a peer as good if it has sent enough blob parts?
See `switch` in `Reactor.peerStatsRoutine()`.

* Add `BlobPartMessage` to WAL replay after crash.

* Fix unit test `TestMarshalJSONPeerState`

* Update unit test `TestMsgToProto`

* Add test `TestBlobPartMessageValidateBasic`

* Small refactor of `State.readReplayMessage()`

* Add a blob ID to some WAL tests

* Address comments from code review

* Register types `BlobPartMessage` and `HasProposalBlobPartMessage` in consensus
reactor `init`()

* Add `TODO` in consensus reactor `subscribeToBroadcastEvents()`.

In this PR we don't call `FireEvent` for `BlobPartMessage`.
We must add this call in a future PR implementing consensus state updates for
blobs. Because this code isn't there yet, there isn't a "correct spot" to place
the call to `FireEvent` yet.